add custom simple auth

+++ src/main/java/ru/kolesnikov/simplechat/controller/UserController.java
@@ -9,8 +9,10 @@ import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.web.bind.annotation.*;
 import ru.kolesnikov.simplechat.controller.dto.UserDTORegistration;
 import ru.kolesnikov.simplechat.controller.dto.UserDTOResponse;
+import ru.kolesnikov.simplechat.exceptions.NotAuthorizedException;
 import ru.kolesnikov.simplechat.kafka.KafkaListeners;
 import ru.kolesnikov.simplechat.model.User;
+import ru.kolesnikov.simplechat.service.AuthService;
 import ru.kolesnikov.simplechat.service.UserService;
 
 import javax.validation.Valid;
@@ -26,10 +28,11 @@ import static ru.kolesnikov.simplechat.kafka.KafkaTopicConfig.KAFKA_TOPIC;
 public class UserController {
 
     private final UserService userService;
-    private final KafkaListeners listeners;
+    private final AuthService authService;
 
-    @GetMapping(value = "/api/v1/user")
-    public List<UserDTOResponse> getAllUsers() {
+    @GetMapping(value = "/api/v1/user/{login}/allusers")
+    public List<UserDTOResponse> getAllUsers(@PathVariable String login) {
+        authService.checkAccess(login);
         return userService.getAllUsers()
                 .stream()
                 .map(user ->
@@ -48,7 +51,8 @@ public class UserController {
                         user.getName(),
                         user.getSurname(),
                         user.getPhotoPath(),
-                        user.getPassword()));
+                        user.getPassword()
+                ));
         return new UserDTOResponse(user.getLogin(),
                 user.getName(),
                 user.getPhotoPath());
@@ -57,6 +61,8 @@ public class UserController {
 
     @GetMapping(value = "/api/v1/user/{login}")
     public UserDTOResponse getUserByLogin(@PathVariable String login) {
+        checkNotAuthorized(login);
+
         User user = userService.findUserByLogin(login);
         return new UserDTOResponse(user.getLogin(),
                 user.getName(),
@@ -66,7 +72,9 @@ public class UserController {
    @PutMapping(value = "/api/v1/user/{login}")
     public UserDTOResponse updateUser(@PathVariable String login,
                                       @RequestBody  @Valid UserDTORegistration user) {
-        userService.updateUser(
+       checkNotAuthorized(login);
+
+       userService.updateUser(
                 login,
                 new User(user.getLogin(),
                         user.getName(),
@@ -81,7 +89,14 @@ public class UserController {
     @DeleteMapping(value = "/api/v1/user/{login}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteUser(@PathVariable String login) {
+        checkNotAuthorized(login);
         userService.deleteUser(login);
+    }
+
+    private void checkNotAuthorized(String login) {
+        if(!authService.checkAccess(login)) {
+            throw new NotAuthorizedException();
+        }
     }
